Revert the IfcGeomObjects API to float precision instead of double. Internally double precision is still used.

+++ src/ifcgeom/IfcGeomObjects.cpp
@@ -50,12 +50,12 @@ bool weld_vertices = true;
 bool convert_back_units = false;
 
 int IfcGeomObjects::IfcMesh::addvert(const gp_XYZ& p) {
-	const double X = convert_back_units ? (double)p.X() / Ifc::LengthUnit : (double)p.X();
-	const double Y = convert_back_units ? (double)p.Y() / Ifc::LengthUnit : (double)p.Y();
-	const double Z = convert_back_units ? (double)p.Z() / Ifc::LengthUnit : (double)p.Z();
+	const float X = convert_back_units ? (float) (p.X() / Ifc::LengthUnit) : (float)p.X();
+	const float Y = convert_back_units ? (float) (p.Y() / Ifc::LengthUnit) : (float)p.Y();
+	const float Z = convert_back_units ? (float) (p.Z() / Ifc::LengthUnit) : (float)p.Z();
 	int i = (int) verts.size() / 3;
 	if ( weld_vertices ) {
-		const VertKey key = VertKey(X,std::pair<double,double>(Y,Z));
+		const VertKey key = VertKey(X,std::pair<float,float>(Y,Z));
 		VertKeyMap::const_iterator it = welds.find(key);
 		if ( it != welds.end() ) return it->second;
 		i = (int) welds.size();
@@ -145,9 +145,9 @@ IfcGeomObjects::IfcMesh::IfcMesh(int i, const IfcGeom::ShapeList& shapes) {
 						gp_Vec normal_direction;
 						prop.Normal(uv.X(),uv.Y(),p,normal_direction);						
 						gp_Dir normal = gp_Dir(normal_direction.XYZ() * rotation_matrix);
-						normals.push_back((double)normal.X());
-						normals.push_back((double)normal.Y());
-						normals.push_back((double)normal.Z());
+						normals.push_back((float)normal.X());
+						normals.push_back((float)normal.Y());
+						normals.push_back((float)normal.Z());
 					}
 				}
 
@@ -167,9 +167,9 @@ IfcGeomObjects::IfcMesh::IfcMesh(int i, const IfcGeom::ShapeList& shapes) {
 					const gp_XYZ v1 = pt2-pt1;
 					const gp_XYZ v2 = pt3-pt2;
 					gp_Dir normal = gp_Dir(v1^v2);
-					normals.push_back((double)normal.X());
-					normals.push_back((double)normal.Y());
-					normals.push_back((double)normal.Z());
+					normals.push_back((float)normal.X());
+					normals.push_back((float)normal.Y());
+					normals.push_back((float)normal.Z());
 					*/
 
 					faces.push_back(dict[n1]);
@@ -198,7 +198,7 @@ IfcGeomObjects::IfcObject::IfcObject(int my_id,
 	// Convert the gp_Trsf into a 4x3 Matrix
 	for( int i = 1; i < 5; ++ i )
 		for ( int j = 1; j < 4; ++ j )
-			matrix.push_back((double)trsf.Value(j,i));
+			matrix.push_back((float)trsf.Value(j,i));
 
 	id = my_id;
 	parent_id = p_id;
